refactor:changes to handle some errors
