+++ src/main/java/com/example/secure_event_hub/exception/GlobalExceptionHandler.java
@@ -1,6 +1,7 @@
 package com.example.secure_event_hub.exception;
 
 
+import com.example.secure_event_hub.controller.ResponseHandler;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -10,31 +11,31 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 public class GlobalExceptionHandler {
 
     @ExceptionHandler(UserAlreadyExistException.class)
-    public ResponseEntity<String> handleUserAlreadyExistException(UserNotFoundException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.CONFLICT);
+    public ResponseEntity<Object> handleUserAlreadyExistException(UserAlreadyExistException exception) {
+        return  ResponseHandler.error(null, exception.getMessage(), HttpStatus.CONFLICT);
     }
 
     @ExceptionHandler(RegistrationFailedException.class)
-    public ResponseEntity<String> handleRegistrationFailedException(RegistrationFailedException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+    public ResponseEntity<Object> handleRegistrationFailedException(RegistrationFailedException exception) {
+        return  ResponseHandler.error(null, exception.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
     }
 
     @ExceptionHandler(InvalidInputException.class)
-    public ResponseEntity<String> handleInvalidInputException(InvalidInputException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.BAD_REQUEST);
+    public ResponseEntity<Object> handleInvalidInputException(InvalidInputException exception) {
+        return  ResponseHandler.error(null, exception.getMessage(), HttpStatus.BAD_REQUEST);
     }
 
     @ExceptionHandler(UserNotFoundException.class)
-    public ResponseEntity<String> handleUserNotFoundException(UserNotFoundException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
+    public ResponseEntity<Object> handleUserNotFoundException(UserNotFoundException exception) {
+        return  ResponseHandler.error(null, exception.getMessage(), HttpStatus.NOT_FOUND);
     }
 
     @ExceptionHandler(InvalidCredentialsException.class)
-    public ResponseEntity<String> handleInvalidCredentialsException(InvalidInputException ex){
-        return  new ResponseEntity<>(ex.getMessage(), HttpStatus.UNAUTHORIZED);
+    public ResponseEntity<Object> handleInvalidCredentialsException(InvalidCredentialsException exception){
+        return   ResponseHandler.error(null, exception.getMessage(), HttpStatus.UNAUTHORIZED);
     }
     @ExceptionHandler(AccountInactiveException.class)
-    public ResponseEntity<String> handleAccountInactiveException(AccountInactiveException ex){
-        return  new ResponseEntity<>(ex.getMessage(), HttpStatus.FORBIDDEN);
+    public ResponseEntity<Object> handleAccountInactiveException(AccountInactiveException exception){
+        return   ResponseHandler.error(null, exception.getMessage(), HttpStatus.FORBIDDEN);
     }
 }
