+++ src/main/java/com/example/secure_event_hub/repository/UserRepository.java
@@ -2,8 +2,16 @@ package com.example.secure_event_hub.repository;
 
 import com.example.secure_event_hub.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
+@Repository
 public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByUserName(String userName);
+
+    Optional<User> findByEmail(String email);
 
 }
